[DsBrowser|DsBrowserSpectrumChart] created spectrum chart modularized component
